backend/requirements.txt
fastapi
uvicorn
spotipy
pandas
numpy
umap-learn
python-dotenv
